return type change for convert

+++ src/Manager/CurrencyExchangeRateManagerInterface.php
@@ -5,7 +5,7 @@ namespace App\Manager;
 interface CurrencyExchangeRateManagerInterface
 {
     // must return result of conversion as float
-    public function convert(string $currencyFrom, string $currencyTo, int $amount): float;
+    public function convert(string $currencyFrom, string $currencyTo, int $amount): int;
 
     // must return an array of currencies with key as the code and value as the name
     public function loadCurrencies(): array;
